# perf: add isSimpleValidator