--- a/lib/helpers/isSimpleValidator.js
+++ b/lib/helpers/isSimpleValidator.js
@@ -22,28 +22,7 @@ function _isSimpleValidator(obj) {
   const keys = Object.keys(obj);
   let result = true;
   for (let i = 0, len = keys.length; i < len; ++i) {
-    if (!result) {
-      break;
-    }
-    if (
-      typeof obj[keys[i]] !== 'object' ||
-      obj[keys[i]] === null
-    ) {
-      continue;
-    }
-    if (Array.isArray(obj[keys[i]])) {
-      for (let j = 0, jl = obj[keys[i]].length; j < jl; ++j) {
-        if (utils.isNativeObject(obj[keys[i]][j])) {
-          continue;
-        }
-        if (typeof obj[keys[i]][j] === 'object' && obj[keys[i]][j] !== null) {
-          result = false;
-          break;
-        }
-      }
-    } else if (utils.isNativeObject(obj)) {
-      continue;
-    } else if (typeof obj[keys[i]] === 'object' && obj[keys[i]] !== null) {
+    if (typeof obj[keys[i]] === 'object' && obj[keys[i]] !== null) {
       result = false;
       break;
     }