--- a/lib/helpers/isSimpleValidator.js
+++ b/lib/helpers/isSimpleValidator.js
@@ -1,7 +1,5 @@
 'use strict';
 
-const utils = require('../utils');
-
 /*!
  * Determines if `arg` is a flat object.
  *