--- a/lib/schematype.js
+++ b/lib/schematype.js
@@ -10,6 +10,7 @@ const $exists = require('./schema/operators/exists');
 const $type = require('./schema/operators/type');
 const handleImmutable = require('./helpers/schematype/handleImmutable');
 const isAsyncFunction = require('./helpers/isAsyncFunction');
+const isSimpleValidator = require('./helpers/isSimpleValidator');
 const immediate = require('./helpers/immediate');
 const schemaTypeSymbol = require('./helpers/symbols').schemaTypeSymbol;
 const utils = require('./utils');
@@ -886,7 +887,7 @@ SchemaType.prototype.validate = function(obj, message, type) {
       properties = { validator: obj, message: message };
       properties.type = type || 'user defined';
     } else if (message instanceof Object && !type) {
-      properties = utils.clone(message);
+      properties = isSimpleValidator(message) ? Object.assign({}, message) : utils.clone(message);
       if (!properties.message) {
         properties.message = properties.msg;
       }
@@ -914,8 +915,8 @@ SchemaType.prototype.validate = function(obj, message, type) {
     arg = arguments[i];
     if (!utils.isPOJO(arg)) {
       const msg = 'Invalid validator. Received (' + typeof arg + ') '
-          + arg
-          + '. See https://mongoosejs.com/docs/api.html#schematype_SchemaType-validate';
+        + arg
+        + '. See https://mongoosejs.com/docs/api.html#schematype_SchemaType-validate';
 
       throw new Error(msg);
     }
@@ -1242,7 +1243,7 @@ SchemaType.prototype.doValidate = function(value, fn, scope, options) {
 
   // Avoid non-object `validators`
   const validators = this.validators.
-    filter(v => v != null && typeof v === 'object');
+    filter(v => typeof v === 'object' && v !== null);
 
   let count = validators.length;
 
@@ -1250,16 +1251,16 @@ SchemaType.prototype.doValidate = function(value, fn, scope, options) {
     return fn(null);
   }
 
-  const _this = this;
-  validators.forEach(function(v) {
+  for (let i = 0, len = validators.length; i < len; ++i) {
     if (err) {
       return;
     }
 
+    const v = validators[i];
     const validator = v.validator;
     let ok;
 
-    const validatorProperties = utils.clone(v);
+    const validatorProperties = isSimpleValidator(v) ? Object.assign({}, v) : utils.clone(v);
     validatorProperties.path = options && options.path ? options.path : path;
     validatorProperties.value = value;
 
@@ -1272,7 +1273,7 @@ SchemaType.prototype.doValidate = function(value, fn, scope, options) {
       return;
     }
 
-    if (value === undefined && validator !== _this.requiredValidator) {
+    if (value === undefined && validator !== this.requiredValidator) {
       validate(true, validatorProperties);
       return;
     }
@@ -1304,7 +1305,7 @@ SchemaType.prototype.doValidate = function(value, fn, scope, options) {
       validate(ok, validatorProperties);
     }
 
-  });
+  }
 
   function validate(ok, validatorProperties) {
     if (err) {
@@ -1374,12 +1375,12 @@ SchemaType.prototype.doValidateSync = function(value, scope, options) {
 
     const v = validators[i];
 
-    if (v == null || typeof v !== 'object') {
+    if (v === null || typeof v !== 'object') {
       continue;
     }
 
     const validator = v.validator;
-    const validatorProperties = utils.clone(v);
+    const validatorProperties = isSimpleValidator(v) ? Object.assign({}, v) : utils.clone(v);
     validatorProperties.path = options && options.path ? options.path : path;
     validatorProperties.value = value;
     let ok = false;
@@ -1454,8 +1455,8 @@ SchemaType._isRef = function(self, value, doc, init) {
       return true;
     }
     if (!Buffer.isBuffer(value) && // buffers are objects too
-        value._bsontype !== 'Binary' // raw binary value from the db
-        && utils.isObject(value) // might have deselected _id in population query
+      value._bsontype !== 'Binary' // raw binary value from the db
+      && utils.isObject(value) // might have deselected _id in population query
     ) {
       return true;
     }
@@ -1496,10 +1497,10 @@ SchemaType.prototype._castRef = function _castRef(value, doc, init) {
   const pop = owner.$populated(path, true);
   let ret = value;
   if (!doc.$__.populated ||
-      !doc.$__.populated[path] ||
-      !doc.$__.populated[path].options ||
-      !doc.$__.populated[path].options.options ||
-      !doc.$__.populated[path].options.options.lean) {
+    !doc.$__.populated[path] ||
+    !doc.$__.populated[path].options ||
+    !doc.$__.populated[path].options.options ||
+    !doc.$__.populated[path].options.options.lean) {
     ret = new pop.options[populateModelSymbol](value);
     ret.$__.wasPopulated = true;
   }