This PR adds a helper method called isSimpleValidator. Basically it is a function to determine if a validator is just a simple object or if it is a more complex one. To avoid the usage in other places, it is not named isFlatObject. It uses a WeakMap to cache the result and gain as such speed.

With isSimpleValidator we determine if we should make a deep clone of the validator or a shallow copy of it with Object.assign.

Before:

```
MONGOOSE_DEV=1 node validate
invalid x 23,461 ops/sec ±0.78% (89 runs sampled)
valid x 75,657 ops/sec ±0.67% (90 runs sampled)
```

after
```
MONGOOSE_DEV=1 node validate
invalid x 29,047 ops/sec ±0.37% (95 runs sampled)
valid x 148,950 ops/sec ±0.29% (93 runs sampled)
```

So we basically double the performance on the valid path.