@Uzlopak 2022-02-17
If we combine this with #11408 we get super performance speed for validation :D