@Uzlopak 2022-02-17
With the improved benchmark from #11415 
before: 

before:
```
MONGOOSE_DEV=1 node validate
invalid async x 21,179 ops/sec ±0.71% (84 runs sampled)
valid async x 41,408 ops/sec ±1.34% (83 runs sampled)
invalid sync x 24,553 ops/sec ±0.76% (93 runs sampled)
valid sync x 76,212 ops/sec ±0.67% (92 runs sampled)
```
after:
```
 MONGOOSE_DEV=1 node validate
invalid async x 23,496 ops/sec ±1.08% (82 runs sampled)
valid async x 60,767 ops/sec ±1.08% (80 runs sampled)
invalid sync x 27,526 ops/sec ±1.27% (92 runs sampled)
valid sync x 142,613 ops/sec ±0.36% (96 runs sampled)
```