@vkarpov15 2022-03-07
I'll merge this and remove the WeakMap caching locally