@Uzlopak 2022-03-07
Seems like the performance is still very good. Strange. Expected a higher performance penalty. But hey, i dont complain.